pyproject.toml
[project]
name = "banking-fraud-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "databricks-sql-connector>=4.2.6",
    "dbt-core>=1.11.9",
    "dbt-databricks>=1.10.9",
    "matplotlib>=3.10.9",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.2",
    "seaborn>=0.13.2",
    "streamlit>=1.57.0",
]
